can't run if you don't have cred file

+++ googleCal.py
@@ -33,24 +33,26 @@ def checkIfGoogleCal():
 def loadGooglCal():
     creds = None
 
-    if os.path.exists("token.json"):
-        creds = Credentials.from_authorized_user_file("token.json", SCOPES)
-    # If there are no (valid) credentials available, let the user log in.
-    if not creds or not creds.valid:
-        if creds and creds.expired and creds.refresh_token:
-            creds.refresh(Request())
-        else:
-            flow = InstalledAppFlow.from_client_secrets_file(
-                f"{os.getcwd()}/creds.json", SCOPES
-            )
-            creds = flow.run_local_server(port=0)
-        # Save the credentials for the next run
-            with open("token.json", "w") as token:
-                token.write(creds.to_json())
+    if os.path.exists("creds.json"):
+        if os.path.exists("token.json"):
+            creds = Credentials.from_authorized_user_file("token.json", SCOPES)
+        # If there are no (valid) credentials available, let the user log in.
 
-    service = build("calendar", "v3", credentials=creds)
+        if not creds or not creds.valid:
+            if creds and creds.expired and creds.refresh_token:
+                creds.refresh(Request())
+            else:
+                flow = InstalledAppFlow.from_client_secrets_file(
+                    f"{os.getcwd()}\creds.json", SCOPES
+                )
+                creds = flow.run_local_server(port=0)
+            # Save the credentials for the next run
+                with open("token.json", "w") as token:
+                    token.write(creds.to_json())
 
-    return service
+        service = build("calendar", "v3", credentials=creds)
+
+        return service
     
 def makeNewAssignment(name, year, month, day, classID):
     with open("setup.json", "r") as f:
@@ -160,3 +162,4 @@ def testEvent(calID):
 
     return error
 
+loadGooglCal()
